requirements.txt
# Syntheverse Core Requirements
# Install with: pip install -r requirements.txt

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# AI/ML
openai>=1.0.0  # For Groq API client

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Blockchain
web3>=6.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0





